refactor: Rename components and directories
