+++ src/containers/ItemsPanel/NoCoursesView/index.jsx
