+++ src/containers/ItemsPanel/NoCoursesView/index.scss
