+++ src/containers/ItemsPanel/NoCoursesView/index.test.jsx
