+++ src/containers/ItemsPanel/index.scss
