+++ src/plugin-slots/ItemsListSlot/images/course_list_slot.png
