+++ src/plugin-slots/NoCoursesViewSlot/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { PluginSlot } from '@openedx/frontend-plugin-framework';
-import NoCoursesView from 'containers/CoursesPanel/NoCoursesView';
+import NoCoursesView from 'containers/ItemsPanel/NoCoursesView';
 
 export const NoCoursesViewSlot = () => (
   <PluginSlot
