+++ src/plugin-slots/README.md
@@ -3,6 +3,6 @@
 * [`org.openedx.frontend.learner_dashboard.course_card_action.v1`](./CourseCardActionSlot/)
 * [`org.openedx.frontend.layout.footer.v1`](./FooterSlot/)
 * [`org.openedx.frontend.learner_dashboard.widget_sidebar.v1`](./WidgetSidebarSlot/)
-* [`org.openedx.frontend.learner_dashboard.course_list.v1`](./CourseListSlot/)
+* [`org.openedx.frontend.learner_dashboard.course_list.v1`](./ItemsListSlot/)
 * [`org.openedx.frontend.learner_dashboard.no_courses_view.v1`](./NoCoursesViewSlot/)
 * [`org.openedx.frontend.learner_dashboard.dashboard_modal.v1`](./DashboardModalSlot)
